LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_opaque_origin.https-expected.txt


FAIL cookieStore in non-sandboxed iframe should not throw assert_equals: cookieStore ${apiCall} should not throw expected "no exception" but got "ReferenceError"
FAIL cookieStore in sandboxed iframe should throw SecurityError assert_equals: cookieStore ${apiCall} should throw SecurityError expected "SecurityError" but got "ReferenceError"
